+++ lib/sorbet-result.rb
@@ -4,8 +4,11 @@
 # frozen_string_literal: true
 
 require "sorbet-runtime"
-require "t/result"
-require "t/nil_payload_error"
-require "t/success"
-require "t/no_payload_on_failure_error"
-require "t/failure"
+require "zeitwerk"
+
+# Sorbet-aware namespace to super-charge your projects
+module Typed; end
+
+loader = Zeitwerk::Loader.new
+loader.push_dir("#{__dir__}/typed", namespace: Typed)
+loader.setup