+++ lib/typed/result.rb
@@ -1,22 +1,22 @@
 # typed: strict
 # frozen_string_literal: true
 
-module T
+module Typed
   # A monad representing either a success or a failure. Contains payload and error information as well.
   class Result
-    extend Sig
-    extend Helpers
-    extend Generic
+    extend T::Sig
+    extend T::Helpers
+    extend T::Generic
 
     abstract!
 
     Payload = type_member
     Error = type_member
 
-    sig { abstract.returns(Boolean) }
+    sig { abstract.returns(T::Boolean) }
     def success?; end
 
-    sig { abstract.returns(Boolean) }
+    sig { abstract.returns(T::Boolean) }
     def failure?; end
 
     sig { abstract.returns(T.nilable(Payload)) }