+++ sorbet/tapioca/require.rb
@@ -1,4 +1,7 @@
-# typed: strict
+# typed: true
 # frozen_string_literal: true
 
-# Add your extra requires here (`bin/tapioca require` can be used to bootstrap this list)
+require "debug"
+require "minitest/autorun"
+require "sorbet-runtime"
+require "zeitwerk"